mmu: elevate memory reads and writes to GameBoy context
The MMU and memory units's read and write methods have been
moved into the GameBoy struct's implementation. This is due to tightly
coupled behavior with the rest of the hardware.
Alternatively, mutable references could be passed around,
but we have found this to make code too verbose and prone to borrow
errors.